.circleci: Prefer using env-file for docker run
'docker run' has the capability to use a file for environment variables,
we should prefer to use that instead of having it be sourced per command
in the docker container.

Also opens the door for cutting down on the total number of commands we
need to echo into a script to then execute as a 'docker exec' command.

Plus side of this approach is that the BASH_ENV is persisted through all
of the steps so there's no need to do any exports / worry about
environment variables not persisting through jobs.

Signed-off-by: Eli Uriegas <eliuriegas@fb.com>

ghstack-source-id: 75c48ecbd0de98eda6d93ef597f39271c5a48f6c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/43293

Signed-off-by: Eli Uriegas <eliuriegas@fb.com>